# scripts/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cadence v2",
    "description": "\n# Multi-Tenant AI Orchestration Platform\n\nProduction-ready framework for deploying AI agent orchestrators at scale.\n\n## Features\n\n* **Multi-Backend Support**: LangGraph, OpenAI Agents SDK, Google ADK\n* **Multi-Tenancy**: Complete tenant isolation with BYOK (Bring Your Own Key)\n* **Three Orchestration Modes**: Supervisor, Coordinator, Handoff\n* **Hot-Reload**: Configuration changes without restart\n* **Scalable Architecture**: Hot/Warm/Cold tier pool (1000+ instances)\n* **Streaming**: Server-Sent Events (SSE) for real-time responses\n\n## Authentication\n\n### JWT Authentication\n```\nAuthorization: Bearer <jwt_token>\n```\n\n### API Key Authentication\n```\nX-API-Key: <api_key>\n```\n\n## Rate Limiting\n\nAll API endpoints are rate-limited per tenant. Default: 100 requests/minute.\n\n## Documentation\n\n* Platform Docs: [README](https://github.com/jonaskahn/cadence)\n* SDK Docs: [cadence-sdk](https://github.com/jonaskahn/cadence-sdk)\n",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.example.com",
      "description": "Staging server"
    }
  ],
  "paths": {
...
    "/api/orgs/{org_id}/completion/stream": {
      "post": {
        "tags": [
          "completion"
        ],
        "summary": "Chat with AI (Streaming)",
        "description": "Send a message to an AI orchestrator and receive a streaming response via Server-Sent Events (SSE).\n\n**Event Types:**\n- `agent_start`: Agent begins processing\n- `agent_end`: Agent completes processing\n- `tool_start`: Tool invocation begins\n- `tool_end`: Tool invocation completes\n- `message`: AI assistant message chunk\n- `error`: Error occurred\n- `metadata`: Additional metadata\n\n**Example SSE Response:**\n```\nevent: agent_start\ndata: {\"agent\": \"supervisor\", \"timestamp\": 1708234567.123}\n\nevent: tool_start\ndata: {\"tool\": \"weather_plugin.get_weather\", \"args\": {\"location\": \"San Francisco\"}}\n\nevent: tool_end\ndata: {\"tool\": \"weather_plugin.get_weather\", \"result\": \"Sunny, 72°F\"}\n\nevent: message\ndata: {\"content\": \"The weather in San Francisco is sunny with a temperature of 72°F.\"}\n\nevent: agent_end\ndata: {\"agent\": \"supervisor\", \"timestamp\": 1708234568.456}\n```",
        "operationId": "chat_stream_api_orgs__org_id__completion_stream_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream of chat events",
            "content": {
              "text/event-stream": {
                "example": "event: message\ndata: {\"content\": \"Hello!\"}\n\n"
              }
            }
          },
          "403": {
            "description": "Access denied to this orchestrator instance"
          },
          "404": {
            "description": "Orchestrator instance not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orgs/{org_id}/completion/chat": {
      "post": {
        "tags": [
          "completion"
        ],
        "summary": "Chat with AI (Synchronous)",
        "description": "Send a message to an AI orchestrator and wait for the complete response.\n\nUse this endpoint when you need the full response at once rather than streaming.\nFor real-time streaming responses, use the streaming endpoint instead.\n\n**Note:** This endpoint will wait until the orchestrator completes processing\nbefore returning. For long-running requests, consider using the streaming endpoint.",
        "operationId": "chat_sync_api_orgs__org_id__completion_chat_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Complete chat response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                },
                "example": {
                  "session_id": "conv_abc123xyz",
                  "response": "The weather in San Francisco is sunny with a temperature of 72°F.",
                  "agent_hops": 3,
                  "current_agent": "weather_plugin"
                }
              }
            }
          },
          "403": {
            "description": "Access denied to this orchestrator instance"
          },
          "404": {
            "description": "Orchestrator instance not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "instance_id": {
            "type": "string",
            "title": "Instance Id",
            "description": "Orchestrator instance identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message",
            "description": "User message",
            "example": "What is the weather in San Francisco?"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id",
            "description": "Conversation ID for maintaining context",
            "example": "conv_abc123xyz"
          }
        },
        "type": "object",
        "required": [
          "instance_id",
          "message"
        ],
        "title": "ChatRequest",
        "description": "Chat request payload.\n\nAttributes:\n    instance_id: Orchestrator instance ID\n    message: User message text\n    conversation_id: Optional conversation ID",
        "example": {
          "conversation_id": "conv_abc123xyz",
          "instance_id": "550e8400-e29b-41d4-a716-446655440000",
          "message": "What is the weather in San Francisco?"
        }
      },
      "ChatResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Conversation session ID",
            "example": "conv_abc123xyz"
          },
          "response": {
            "type": "string",
            "title": "Response",
            "description": "AI assistant response",
            "example": "The weather in San Francisco is sunny with a temperature of 72°F."
          },
          "agent_hops": {
            "type": "integer",
            "title": "Agent Hops",
            "description": "Number of agent routing hops",
            "example": 3
          },
          "current_agent": {
            "type": "string",
            "title": "Current Agent",
            "description": "Last agent that processed the request",
            "example": "weather_plugin"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "response",
          "agent_hops",
          "current_agent"
        ],
        "title": "ChatResponse",
        "description": "Chat response payload for synchronous endpoint.\n\nAttributes:\n    session_id: Conversation session ID\n    response: AI response text\n    agent_hops: Number of agent hops executed\n    current_agent: Last agent that processed the request",
        "example": {
          "agent_hops": 3,
          "current_agent": "weather_plugin",
          "response": "The weather in San Francisco is sunny with a temperature of 72°F.",
          "session_id": "conv_abc123xyz"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from authentication endpoint. Format: `Bearer <token>`"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication. Format: `X-API-Key: <key>`"
      }
    }
  },
...
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
...
}